Rename map.py to map_data.py
