+++ utils/__init__.py
@@ -5,7 +5,7 @@ from .coordinate import Coordinate
 from .counter import Counter
 from .directions import Directions
 from .icon import Icon
-from .map import MapData
+from .map_data import MapData
 from .tile import Tile
 
 # static constants that others may need
